feat(schemas): add UpdateCredentialsSchema
- Added new schema `UpdateCredentialsSchema` to handle user credentials updates in users.ts
- Removed `UpdateCredentialsSchema` from auth.ts for separation of concerns and better organization of schemas.
