+++ src/schemas/auth.ts
@@ -10,10 +10,3 @@ export interface Credentials extends Static<typeof CredentialsSchema> {}
 export interface Auth extends Omit<Credentials, 'password'> {
   roles: string[]
 }
-
-export const UpdateCredentialsSchema = Type.Object({
-    username: Type.String(),
-    password: Type.String(),
-    newUsername: Type.Optional(Type.String()),
-    newPassword: Type.Optional(Type.String())
-})
