+++ src/schemas/users.ts
@@ -1,0 +1,8 @@
+import { Type } from '@sinclair/typebox'
+
+export const UpdateCredentialsSchema = Type.Object({
+  username: Type.String(),
+  password: Type.String(),
+  newUsername: Type.Optional(Type.String()),
+  newPassword: Type.Optional(Type.String())
+})
